Now goals are reloaded when the level is cleared
playerController system has the logic for this

In order to test this logic, a new "secret level" and its respective
goal were added to the entitiesData table of the "test" level table.

When the player contacts a goal, all the goal entities in the current stage
are removed. The goal entities data is scanned for new goals
corresponding to the current level, and if there are some, they are
added to the main componentsTable. Note that this logic appears both
control.lua and main.lua, maybe some refactoring could get a better
or, at least, more elegant approach.